move _common_test.py from tritonsrc/ to test/
